requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
airtable-python-wrapper
pydantic[dotenv]
pydantic[email]
